+++ test/runtests.jl
@@ -12,6 +12,22 @@ using Test
         @info "Testing constructors..."
         include("test_constructors.jl")
     end
+    @testset "Analytic" begin
+        @info "Testing analyzers against closed-form solutions..."
+        include("test_analytic.jl")
+    end
+    @testset "Forward passes" begin
+        @info "Testing number of forward passes..."
+        include("test_forward_passes.jl")
+    end
+    @testset "Enzyme" begin
+        @info "Testing analyzers using Enzyme..."
+        include("test_enzyme.jl")
+    end
+    @testset "AD backends" begin
+        @info "Testing analyzers on several AD backends..."
+        include("test_backends.jl")
+    end
     @testset "CNN" begin
         @info "Testing analyzers on CNN..."
         include("test_cnn.jl")